kvserver: move checkForcedErr to kvserverbase
This is a prereq to making log application stand-alone, since applying
entries means calling this method to determine whether to apply an
empty entry instead.

Touches https://github.com/cockroachdb/cockroach/issues/75729.

Epic: CRDB-220
Release note: None
