+++ pkg/kv/kvserver/kvserverbase/BUILD.bazel
@@ -21,6 +21,7 @@ go_library(
         "//pkg/settings",
         "//pkg/util/errorutil",
         "//pkg/util/hlc",
+        "//pkg/util/log",
         "//pkg/util/quotapool",
         "//pkg/util/tracing/tracingpb",
         "@com_github_cockroachdb_errors//:errors",
